+++ src/common/js/index.js
@@ -7,3 +7,4 @@ import './constants.js';
 import './i18n.js';
 //
 import './configure.js';
+import './functions.js';