+++ ChangeLog.md
@@ -2,6 +2,13 @@
 
 ## ChangeLog
 
+### 1.1.0-rc.0
+
+    Release date: 
+
+    - Define new AccountsZimbra.authenticate() common function
+    - Define new AccountsZimbra.isZimbraUser() common function
+
 ### 1.0.0
 
     Release date: 2026- 4-18